Tons of improvs

+++ src/config.zig
@@ -1,9 +1,9 @@
 // Generate map based on this size
 pub export const mapWidth: u8 = 100;
 pub export const mapHeight: u8 = 60;
-pub const mapSize: usize = @as(u16, mapHeight) * @as(u16, mapWidth);
+pub const mapSize: u16 = @as(u16, mapHeight) * @as(u16, mapWidth);
 
 // ViewPort sizes
 pub export const viewWidth: u8 = 28;
 pub export const viewHeight: u8 = 16;
-pub const viewSize: usize = @as(u16, viewHeight) * @as(u16, viewWidth);
+pub const viewSize: u16 = @as(u16, viewHeight) * @as(u16, viewWidth);
